// src/enums/raca-enum.ts
/* eslint-disable no-unused-vars */

export enum RacaEnum {
  anao = 'anao',
  anfibio = 'anfibio',
  centauro = 'centauro',
  construto = 'construto',
  elfo = 'elfo',
  esqueleto = 'esqueleto',
  goblin = 'goblin',
  halfling = 'halfling',
  humano = 'humano',
  licantropo = 'licantropo',
  meio_elfo = 'meio_elfo',
  meio_orc = 'meio_orc'
}